# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Restful API for web app to buy air tickets
  version: 1.1.0
servers:
  - url: https://api.airtickets.com
paths:
...
  /tickets:
    post:
      summary: Create ticket
      tags:
        - Tickets
      operationId: createTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ticket"
      responses:
        '200':
          description: Successful response with created ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tickets"
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
  /flights/{flightId}/tickets:
    get:
      summary: Get tickets on a flight
      tags:
        - Tickets
      operationId: getTicketsByFlightId
      parameters:
        - name: flightId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tickets"
        '401':
          description: User is unauthorized
  /tickets/{ticketId}:
    patch:
      summary: Modify ticket by id
      tags:
        - Tickets
      operationId: changeTicketById
      parameters: 
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ticket"
      responses:
        '200':
          description: Successful response with modified ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
        '404':
          description: Ticket is not in DB
    delete:
      summary: Delete ticket by id
      tags:
        - Tickets
      operationId: deleteTicketById
      parameters: 
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '401':
          description: User is unauthorized
        '404':
          description: Ticket is not in DB
...
  /services:
    get:
      summary: Get all services
      tags:
        - Services
      operationId: getAllServices
      responses:
        '200':
          description: Successful response with a list of services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Services"
    post:
      summary: Create service
      tags:
        - Services
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Service"
      responses:
        '200':
          description: Successful response with created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        '204':
          description: Service exists
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
  /services/{serviceId}:
    patch:
      summary: Modify service by id
      tags:
        - Services
      operationId: changeServiceById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Service"
      parameters: 
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with modified service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        '400':
          description: Incorrect input data
        '401':
          description: User is unauthorized
        '404':
          description: No services added for this ticket
    delete:
      summary: Delete service by id
      tags:
        - Services
      operationId: deleteServiceById
      parameters: 
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '401':
          description: User is unauthorized
        '404':
          description: Service is not it DB
  /tickets/{ticketId}/services:
    get:
      summary: Get service by ticketId
      tags:
        - Services
      operationId: getServiceByTicketId
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Services"
        '401':
          description: User is unauthorized
  /tickets/{ticketId}/services/{serviceId}:
    delete:
      summary: Delete service by ticketId
      tags:
        - Services
      operationId: deleteServiceByTicketId
      parameters: 
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
        '401':
          description: User is unauthorized
        '404':
          description: Service is not in DB
  /orders/{orderId}/tickets:
    post:
      summary: Add ticket to order
      tags:
        - Tickets
      operationId: orderTicket
      parameters: 
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
        - name: ticketId
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response with an ordered ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        '401':
          description: User is unauthorized
...
components:
  schemas:
...
    Ticket:
      type: object
      required:
        - ticketId
        - flightId
        - row
        - place
        - class
        - refund
      properties:
        ticketId:
          readOnly: true
          type: integer
          example: 2
        flightId:
          type: integer
          example: 34
        orderId:
          readOnly: true
          type: integer
          example: 5
        row:
          type: integer
          example: 3
        place:
          type: string
          example: B
        class:
          type: string
          enum:
            - эконом
            - бизнес
            - первый
          example: бизнес
        refund:
          type: boolean
          example: true
    Tickets:
      type: array
      items:
        $ref: "#/components/schemas/Ticket"
...
    Service:
      type: object
      required: 
        - serviceId
        - name
        - price
        - economyClass
        - businessClass
        - firstClass
      properties:
        serviceId:
          readOnly: true
          type: integer
          example: 3
        name:
          type: string
          example: Бронирование отеля
        price:
          type: string
          example: 5000
        economyClass:
          type: boolean
          example: false
        businessClass:
          type: boolean
          example: false
        firstClass:
          type: boolean
          example: true
    Services:
      type: array
      items:
        $ref: "#/components/schemas/Service"
...